refactor: 로또 생성 팩토리 제거하여 lotto creator를 담당하는 util성 객체 생성
