--- src/main/java/lotto/model/LottoFactory.java
@@ -1,8 +1,0 @@
-package lotto.model;
-
-import java.util.List;
-
-public interface LottoFactory {
-
-    List<Lotto> create(int purchasedCount);
-}
