+++ src/test/java/lotto/model/LotteriesTest.java
@@ -4,7 +4,6 @@ import lotto.model.enumeration.Rank;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -12,26 +11,6 @@ import java.util.stream.Collectors;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class LotteriesTest {
-
-    @Test
-    public void 로또_당첨금_검증() {
-        Map<Rank, Long> lotteriesRank = Map.of(Rank.FIRST, 2L, Rank.THIRD, 1L);
-
-        Lotteries lotteries = createLotteries(3000);
-
-        long result = lotteries.getTotalWinningMoney(lotteriesRank);
-
-        assertThat(result).isEqualTo(4001500000L);
-    }
-
-    @Test
-    public void 로또_당첨금_없는_검증() {
-        Lotteries lotteries = createLotteries(3000);
-
-        long result = lotteries.getTotalWinningMoney(new HashMap<>());
-
-        assertThat(result).isEqualTo(0);
-    }
 
     @Test
     public void 로또_당첨_검증() {
@@ -42,7 +21,7 @@ class LotteriesTest {
 
         List<LottoNumber> lastWinLotto = createLottoNumber(1, 2, 4, 24, 25, 45);
 
-        Map<Rank, Long> result = lotteries.getLotteriesRank(lastWinLotto, 33);
+        Map<Rank, Long> result = lotteries.getLotteriesRank(new Lotto(lastWinLotto), 33);
 
         assertThat(result).isEqualTo(Map.of(Rank.FIRST, 1L, Rank.FIFTH, 1L));
     }
@@ -56,13 +35,9 @@ class LotteriesTest {
 
         List<LottoNumber> lastWinLotto = createLottoNumber(1, 2, 4, 24, 25, 43);
 
-        Map<Rank, Long> result = lotteries.getLotteriesRank(lastWinLotto, 45);
+        Map<Rank, Long> result = lotteries.getLotteriesRank(new Lotto(lastWinLotto), 45);
 
         assertThat(result).isEqualTo(Map.of(Rank.SECOND, 1L, Rank.FIFTH, 1L));
-    }
-
-    private Lotteries createLotteries(int amount) {
-        return new Lotteries(new LottoMachine(), amount);
     }
 
     private List<LottoNumber> createLottoNumber(int... number) {
